codforce round 472 b, d modify

+++ codeforce/Round_472/div2_d.cpp
@@ -7,7 +7,7 @@ int n, m[100002];
 int ans[100002][2];
 
 int main(){
-  int ans2 = 0;
+  long long ans2 = 0;
 
   cin >> n;
   for(int i = 1; i <= n; i++)
@@ -28,7 +28,7 @@ int main(){
     ans[i - 1][0] = max(ans[i - 1][0], ans[i][0] - 1);
   }
 
-  printf("%d\n", ans2);
+  printf("%lld\n", ans2);
 
   return 0;
 }
